Store IKE setup data in states instead of as global variables
This commit creates three classes to carry data that is only used
during IKE INIT or IKE AUTH exchange. Those data is usually created
in one state and need to be passed to the later state(s). Instead of
storing those data as global variables, this commit update the
IkeSessionStateMachine to pass them to next state by directly
setting the next state with them.

Bug: 193145138
Test: atest FrameworksIkeTests, CtsIkeTestCases
Change-Id: I5a329b05859ffc1d92ba87e598e2092c1682bbfd