Support Cts test: INITIAL CONTACT Notify payload
This commit indicates Cts test case delivery for
validating INITIAL CONTACT Ike Option
functionality for the IKE Session.

If this option is set for an IKE Session, The
INITIAL CONTACT  message is sent to the peer
to indicate this IKE SA is the only IKE SA
currently active between the authenticated
identities.

Bug: 186179788
Test: atest CtsIkeTestCases(new tests added)
Change-Id: Ia1a8b2c4907dc2a77cfdfc9a14cb11b183e892a7