copy files in global config during remote configuration
Test host global config contains references to files such as
service account keys. Right now those references are set in the
remote global config with the same value as is, but the files
are never transfered over. It currently works in GCE usage
possibly because the invalid file names are ignored and hosts
are in fact using GCE default machine service account.

This change adds the ability to modify targeted file type
option values while cloning for remote invocaiton, and at
the same time, the affected files are collected in a mapping
from local file path to the future remote path. Lastly the files
are transferred over scp into their locations in the remote tf
invocation folder (and eventually deleted after completion as
part of the teardown).

Bug: 202400161
Test: added and existing unit tests, e2e validation
Change-Id: Id4add705fd4e79448a7b469193d28ca8bde809b9