Add an exception class in the device utils class
Previously the launchPackage method returns false when the command failed, which means the caller does not have any more information about the failure other than the fact that the operation failed. Test failure message won't contain enough information about the failure and the users will need to look into the logs to find the cause.

Test: atest csuite-harness-tests
Fix: 205912019
Change-Id: I1d9a0b4ea03049e4ff88b983a508a41ce8a71a62