Autoupdate of 2 submodules
*external/elfutils updated to eed13d52bc4865ee18a257970be91b1402c7833b
Merge "Extend visibility to //external/stg"
*external/libabigail updated to 42c7a4e0f3de01e85a2fe2bb513c77d8316a64e0
Extend (header) visibility to //external/stg

This extends the general visibility to //external/stg.
Further, `libabigail` does now also export headers to targets that link
`libabigail` as static lib. This was not needed before because of the
lack of external library users.
stg is currently using libabigail headers that are not part of the
public interface (mostly due to unneccessary exposure), but this is
generally fine. Hence add `libabigail_src_headers` to expose those
internal headers specifically to //external/stg only. This will link as
we do not apply any symbol visibility restrictions to the built
libabigail (as opposed to the upstream libabigail build).

Bug: 188176217
Signed-off-by: Matthias Maennich <maennich@google.com>
Change-Id: I6022906a4d6cf9ce315b0ebe51ea1b7489f69ca6