Fix wrong package availability check for multi-user.
Currently we are checking if the ICS package is enabled during runtime
before we try to bind to it. However, if we switched to second user,
checking availability will always return false. Fix this by checking
it with package manager of the current user.

Test: make sure Telecom can bind to all non-UI ICS after switched to
second user manually
Bug: 197820445

Change-Id: I7944705fff1323c471a378c3dd7ee3cd8911dbc0
Merged-In: I7944705fff1323c471a378c3dd7ee3cd8911dbc0