Fix Elf caching.
There were a few places where Elf caching didn't work when trying to
cache apks. There were also a few places where caching of a read-only
and read-execute map that represents a single elf wasn't working
properly either. In addition, the elf_start_offset and elf_offset
was not getting set properly in many cases.

Fix these and change the caching strategy from file:offset to a
map of maps, first map is the filename, the second is the elf_start_offset.

Replace all of the elf cache tests with better tests that actually
test all of the varied paths.

There is also a potential bug where the Elf object was being modified
outside of the MapInfo lock. Fixed by making sure the modification is
set within the lock.

Bug: 193408481
Test: Unit tests pass.
Change-Id: I517ae316151b2b608afa62a5863cf632752f29b2