Enable building non-snapshot packages.
This used to not matter because we created the POM and AAR ourselves,
but now that we're letting gradle do the work we need to not build
snapshot packages for releases so we get the correct artifact names.

I added a new project property ("release") that controls whether or
not the build is a snapshot or not, and split the release build
scripts in two so the release build can set that without setting it
for pre/post submit.

The regular build.sh is sticking around transitionally while I fix the
build configs.

Test: ./scripts/build_release.sh
Test: ./scripts/build_snapshot.sh
Bug: None
Change-Id: I220f265fafabc7c0ca3cdeb5321b65a59c52858a