Autoupdate of 3 submodules
*cts updated to 8a328df4c0406e9a256a8ff86156075c12c2c887
Merge "Assert that the @FixedSize AIDL union is always the same size"
*system/libfmq updated to 75253c2507808aaa9fed7720b20fee221580d0e6
Added an example of @FixedSize union

Bug: 206496560
Test: atest fmq_test fmq_unit_test
Change-Id: I1343356e9bd41d75c6a71fa970d6b1dc5e216f3f

*system/tools/aidl updated to 97be5845253b8a16d3448bd8cad2dd23b156732c
Add @FixedSize union

@FixedSize union uses a custom implementation rather than std::variant
so that its value can have a stable/fixed layout and be passed via FMQ.

A custom implementation relies on the fact that @FixedSize
union/parcelable should have "trivially copyable" fields. So we can't
use it for union in general.

It would be nice if we could replace std::variant with it and have the
same implementation for union types. Maybe in the future.

Bug: 206496560
Test: golden_test.sh check
Test: aidl_integration_test (no regression)
      atest fmq_test fmq_unit_test
Change-Id: I86351ecb1990818d0fe2faaad7c4b8c2c33881b5