Autoupdate of 2 submodules
*hardware/interfaces updated to 71b595549cdd1b12dfd699eb079ae920f200696e
Merge "Make NNAPI Version more structured -- hal"
*packages/modules/NeuralNetworks updated to cbc13302f4be78580f6eb4bfe03bf4abd9f545d8
Make NNAPI Version more structured -- runtime

Prior to this change, NNAPI Versions were linear and represented by an
enumeration. However, this did not properly account for the
non-linearity of runtime-specific features such as a control flow
operations with operands of dynamic sizes. This change alters Version to
be a struct containing a feature level enumeration as well as a boolean
which indicates whether there are runtime-specific features.

Bug: 206975939
Test: mma
Test: NeuralNetworksTests_static
Change-Id: Id71bb51d63c233d0d6017f14b471a459acbdc944