Support Passpoint in CaptiveportalLogin
A passpoint WiFi AP is detected as a captive portal. The login
process requires to download a certificate and install the
certificate. CaptivePortalLogin currently support files download
but did not support the passpoint file type. The certificate
will be treated as a normal unknown file. User may need to
manually install it or do the flow via other browsers. Add the
wifi passpoint config file support and trigger wifi certificate
installer to support passpoint in this scenario

Bug: 179514300
Test: atest CaptivePortalLoginTest
Test: manaually test with downloading x-wifi-config file
Change-Id: I0a21da0dbfd0f48f3c393b860a2e054c6e3dfa35