Increase LUCI build timeout to avoid fugu INFRA_FAILURE
One of our fugu devices failed and we have no replacement for now.

This means we have one fugu device handling two distinct builds,
which makes LUCI infra-fail because the builds are not started
in a timely manner.  Increase the timeout to mitigate this.

When a new CL is submitted, this means that one of the fugu builds
will take ~11 hours to even start, let alone finish (hence the
current complains from LUCI), but let's tolerate it for now.

Test: ./validate.sh
Change-Id: Ia1d50b1a4c415310544ee7bee146c495e5b28b88