Read stdout from llvm-symbolizer during native stack dump.
We initially run llvm-symbolizer just to test if it is available
and functioning.  However, we run with it own stdout pipe,
which we promptly close before it even starts executing.

Properly consume the stdout output instead, to stop the tool spamming
stderr with (justified) complains. (however, other than the log spam,
this had no other negative consequences on symbolization)

Test: check logcat of crashing test; the extra errors are gone
Change-Id: Id96f8a3202976ea7208860e131e8eddb4beae612