Autoupdate of 2 submodules
*toolchain/android_rust updated to 08b2b05d757654770e4aeb3bd44c353eaa730686
Update the patch files in preparation for Rust 1.57.0

This CL removes, updates, and adds patches.  The removed patch
stated that we wanted to use opt level 2 instead of 3 because LLVM 7.0
would segfault if it was compiled with opt level 3.  As Rust now uses
LLVM 13 and the compiler was able to bootstrap itself without this patch
it seems safe to remove it.  Further testing will be done with generated
Android system images.

The added patch is a cherry-pick of a patch that will be in the Rust
1.58 release.  It converts a log message concerning missing metadata
from a warning to an info message, thus cleaning up our build logs.
This is necessary because some of our Rust crates link against C/C++
shared libraries that do not contain the `.rustc` sections that are
supposed to contain Rust metadata.

Bug: 203802952
Test: ./build.py --lto thin
Test: Build cuttlefish image and boot it
Change-Id: I9d86bdfa26bf64c1a022b84e94b004d99da6c115

*toolchain/rustc updated to a1538426d0406135b8a525ab5658419cd2a5ba4e
Importing rustc-1.57.0

Bug: 203802952
Test: ./build.py --lto=thin
Change-Id: I45a7d6b548bf423bac860f01bd5ad978a95fe6df