Autoupdate of 2 submodules
*external/rust/crates/anyhow updated to f0fec759813f457b94eda525acba5afe67cafb22
Support RUST_BACKTRACE in test_fmt

Setting `RUST_BACKTRACE` when running tests is useful so that the logs
will contain backtraces when a test fails. `anyhow` uses this
environment variable as well when formatting its errors. `test_fmt`
currently tests that formatting is an exact match, so adding a trailing
backtrace causes them to fail.

This patch checks that the formatted error *starts with* the intended
result, thus allowing it to pass when a backtrace is appended.

Test: atest --host anyhow_host_test_tests_test_fmt w/RUST_BACKTRACE=full
Bug: 193841423
Change-Id: If52791f53e0deefea815c24cc4f7695201de7d79

*tools/tradefederation/core updated to a0f94e5c82cfb27b4ba6e7bd9bdbcd403b100357
rust: Enable backtraces on failing tests

Set RUST_BACKTRACE=full when running Rust tests so that failures in CI will
include the backtrace.

Test: Use atest --host to run an intentionally broken test, observe backtrace
Bug: 193841423
Change-Id: I0f1e706292b186c16ca0cbb0a9963f1a61909fd0