Autoupdate of 2 submodules
*frameworks/base updated to 5f5ff758707da0fc17fdcf03902d27d337d377d8
Merge "Add @JavaOnlyStableParcelable to AttributionSource"
*system/tools/aidl updated to 30c8e5270fc3edb18754c9c628fecc08c511e8f7
Use PermissionChecker for Java backend

Move to PermissionChecker.checkPermission to validate the permissions.
Use the ForDataDelivery variant as default (the Preflight variant can be
added later on, using a different predicate for example).

If the method receives an AttributionSource as parameter, it will be
automatically used for the permission check. Otherwise, a new
AttributionSource will be generated.

Ran bpfmt on Android.bp, some lines have been reformatted.

Bug: 197828948
Bug: 208707422
Test: atest --host aidl_unittests
Change-Id: I0ca010eafba0fc02a614f1add9275dd92938d623