Migrate PATCHES.json to new metadata schema
In an effort to automate patch synchronization between CrOS and
Android LLVM, we're moving the "comment" and "bugs_tests" fields
in PATCHES.json into a "metadata" dictionary entry instead.

Additionally, this sorts the key fields in PATCHES.json so that
keys won't be shuffled around as we apply new patches from different
programs.

Test: N/A
Bug: http://b/209493133
Change-Id: If8824d572bacc789529a65e15f9fe5f71ed64755