Reopen autcomplete panel after a device spec key is selected
When a user selects a device spec key on test_run_target_picker, the key is appended to the input field, and the user needs to input the value. By default, the autocomplete component closes the panel after an option is selected. This CL adds the condition that reopens the panel.

PiperOrigin-RevId: 415151653
Change-Id: Ib994d532282e970a1ce5483222d93a5c5b0509a5