Updating handling of LTO related linker flags
Previously, b/201551165 prevented the use of LTO with the
armv7-linux-androideabi target due to symbol visibility issues.  This CL
restructures how some of the linker arguments are passed and uses this
functionality to get around the symbol visibility issue.  As a result
the code is cleaner and LTO is now enabled for armv7.

Bug: 201551165
Test: ./build.py --lto=thin
Change-Id: I1a435d549ab11f4893dd699b3ded78834257355d