Add VNDK namespace to Vendor/Product APEX section
VNDK namespace would not be added to Vendor/Product APEX section if the
APEX is marked as 'use_vndk_as_stable' is false, but still there are
some namespaces require for VNDK namespace. This change adds VNDK
namespace in the section by default, so any namespaces can refer to the
namespace if needed.

Bug: 209491899
Test: CF boot succeeded and did not show linker error with Vendor APEX
Change-Id: I9fc6864d943a498d012d8ef8c8519da1180cc58d