Add SynchronousResultReceiver to modules
It aim to be used in Binder call, to easily add a oneway interface on
existing api.

FYI: This is a re-used of existing sync receiver:
f/base/.../com/android/internal/util/SyncResultReceiver.java
f/base/.../android/os/SynchronousResultReceiver.java

The first one is in internal and can not be open to modules.
The second one is in android.os but we feel like it should not be moved
to a systemApi there. This class introduce some async behavior and the
api council seems to not like theses
Bug: 200200870
Tag: #refactor
Test: Build - start BT - play music
Test: atest ModulesUtilsTests

Change-Id: I15d72e2538f00ff8ea77d3c58a7b3ead11e7965c