Log additional details in NativeDevice#logDeviceActionException
NativeDevice#performDeviceAction currently consumes exceptions caught
from action#run with only the Exception class and message logged; this
can make debug of host side tests difficult if a generic IOException
is the cause of a test failure. This commit adds additional logging
to help identify the source of a caught exception.

Bug: 203455834
Test: atest ListeningPortsTest with EOFException thrown from ddmlib
Change-Id: I517c8be258d3fb8698ea2b5091c7d7eb0a78ae64