Add fmtlib_headers to the allowed list.
fmtlib has been in the list. The header variant of the lib was created
and it needs to be added as well.

Bug: 209929099
Test: m
Test-Info: No behavior change. fmtlib has been in the allow list. The
header variant of the lib was created and it is used in the header-only
contexts.
Apex-Size-Increase: 0
Previous-Platform-Support: 29 (the same as min_sdk_version of fmtlib)
Aosp-First: Yes.

Change-Id: Ic430285fdb2d64077e6d61c81c78aab4a9842ed7