Dexopt service-wifi.
`service-wifi` is a jar that system_server loads dynamically. Such jars
can be dexopted now. This CL adds `service-wifi` to a
`systemserverclasspath_fragment`, which does not actually add the jar to
the classpath, but tells the build system to dexopt it.

NOTE: The corresponding `exported_systemserverclasspath_fragments` needs
to be added to the `prebuilt_apex` once a new prebuilt is dropped.

Bug: 203198541
Test: m com.android.wifi
Change-Id: Iaf045fa8b320ab631c59a6f4a5165b8c3a7a7d13