Dexopt service-scheduling.
`service-scheduling` is a jar that system_server loads dynamically. Such
jars can be dexopted now. This CL adds `service-scheduling` to a
`systemserverclasspath_fragment`, which does not actually add the jar to
the classpath, but tells the build system to dexopt it.

NOTE: The corresponding `exported_systemserverclasspath_fragments` needs
to be added to the `prebuilt_apex` once a new prebuilt is dropped.

Bug: 203198541
Test: m com.android.scheduling
Change-Id: Ie851c6a2d4f1f890f62dd6db52e692ef1e615da2