Only post viewport command on GL.
The viewport command on Vulkan is a NOP, save for rebinding
the subwindow, which will happen on any other swapchain-touching
post command.

This is a quick fix for a deadlock issue that is causing
b/209561395. While explained in more detail on the bug, the high-
level sumamry is that three threads (PostWorker, virtio_gpu, and
gpu_display_*) can interweave in such a way that they all deadlock
due to a locked sVkEmulationLock and a locked Framebuffer lock.

The real fix for this deadlock issue is going to require some more
thought.

Bug: 209561395
Test: Could not reproduce issue after this fix. Note that this issue
is not always reproducible for me even without the fix.

Change-Id: I4be08bc851330cc9ecc55a925f61fd2686c345fa