Temporary workaround for relaying probe_req frame at wmediumd
Wmediumd does not redirect probe_req frames even if it is broadcasting
frame because some field of probe_req(tx_rates) is invalid. It seems
that is needs to modify mac80211_hwsim driver to fill that field. As
temporary solution, fill tx_rates when frame is probe_req and tx_rates
is empty. Will remove this when change of mac80211_hwsim is applied.

Bug: 199351534
Bug: 211353765
Test: lunch aosp_cf_x86_64_phone-userdebug && m \
        && launch_cvd
      Turn on Wi-Fi and check scanned AP list doesn't disappears
Test: Run ATP test suite/test-mapping-presubmit-large-retry_cloud-tf
      and check CtsWifiTestCases doesn't fails

Change-Id: I1f1eb6a19b9f4f9967c3bbfdf079bf8bc42aa4e3