Autoupdate of 2 submodules
*art updated to e7815b8c6553171d1bace73bcc5c9d2b9b08de29
Add a dex2oat flag to optimize apps for JIT Zygote environment.

This change allows us to control the boot image to use within the ART
module.

It is a no-op change for now as the behavior of the new flag mirrors the
current behavior of installd.

Also: allow missing primary boot image for app compilation.

Bug: 203492478
Test: atest art_standalone_dex2oat_tests
Test: manual -
  1. Build a system image and flash it into a device.
  2. adb shell setprop dalvik.vm.profilebootclasspath true
  3. adb shell pm compile -m speed -f com.android.htmlviewer
  4. See dex2oat running with the primary boot image only
     (http://gpaste/5031926999023616)
Change-Id: I760808302d4760330befc271c9312018ed1671ef

*frameworks/native updated to 9a869c9c1f78e8b8c19ff4f04e6eaf0a70e094c2
Merge "Update JIT Zygote image option for installd and otapreopt."