minijail: Support octal number values in python parser
Currently we use `int()` in the python parser to convert strings to
integers, but it actually behaves differently with `strtol()` in C, the
function used in minijail, specifically for octal numbers.

To resolve this, instead of using `int()` directly, we intimate the
behavior of `strtol` by numberating all the radices supported and
determining the value by regex in the python parser.

Bug: 212217744
Test: (cd tools/ && python3 -m unittest *_unittest.py)
Test: (cr) for f in /build/${BOARD}/usr/share/policy/*.policy; do python
~/chromiumos/src/aosp/external/minijail/tools/compile_seccomp_policy.py
--arch-json "/build/${BOARD}/build/share/constants.json" $f /dev/null;
done

Change-Id: Ie56b0ee3f12c9b937dd584744e0464ed8b9e88a4