Add to check if device has metadata partition
If there is no metadata partition, the tests for metadata partition should
not be suitalbe for that device. metadata partition will be used for feature
Userdata Check Point(UDC) and metadata encryption.
If the device does not support both of them, metadata partition is completely
no meaning and the tests should be skipped.

Bug: b/175853976
Change-Id: Ibd789bcd0482b6af6a0e8b755e93c5ea560ee957
Test: run vts -m VtsGsiBootTest -t VtsGsiBootTest#MetadataPartition.MinimumSize_32bit
Test: run vts -m VtsGsiBootTest -t VtsGsiBootTest#MetadataPartition.FirstStageMount_32bit