DoH: Change to use runtime with current thread scheduler in doh_query()
Perfetto shows that using runtime with multi-thread scheduler
in doh_query() results in 8 threads created and terminated every time
doh_query() is called, which takes approximately 2 ms. So,
change the code to use a runtime with current thread scheduler.
This approximately reduces the time overhead of doh_query()
from 2 ms to 0.5 ms.

Bug: 212364676
Test: cd packages/modules/DnsResolver && atest
Change-Id: I7652e6982b1485ee435b6399798b19badd264500