ANDROID: Preload partitions in close proximity to save RAM
Previously, boot, vendor_boot, and init_boot partitions were preloaded
with enough space (max 64MB which corresponds to their maximum size)
between them. As a result, the minimum RAM requirement was 128MB +
sizeof(init_boot) even when the three partitions are considerably
smaller than that.

This CL fixes the issue by preloading them next to each other - no gap
between them - IF the three can fit in the first 64MB region. In this
case, the safe distancing is not necessary because contents from
vendor_boot and init_boot will be relocated to the place AFTER the 64MB
region and nothing will be copied into the place where the two
partitions are preloaded. So there is no fear of overwriting during the
relocation.

Note: when their sum exceeds 64MB, we keep the existing behavior. This
is because it is not trivial (though not impossible) to do the
relocation without overwriting.

Bug: N/A
Test: boot microdroid

Change-Id: I678a90cf6810938db26285998ec3f4532a764760