Ignore changes to irrelevant APEXes
When reading the list of APEXes, ignore any APEXes that do not contain
any compilable JARs. Changes to such APEXes now do not and should not
trigger recompilation.

This is helpful for the CompOS use case, since it doesn't even see
these APEXes.

As part of this, extract existing code to determine APEX name from
location and make it resuable.

Bug: 210473615
Test: atest art_standalone_libartbase_tests
Test: Presubmits
Change-Id: I3f64e6228f091cabce33a292058f960787780f62