Autoupdate of 3 submodules
*cts updated to cd5b8fb76ff855fd071d2466874fba87499e28ef
Merge "Add EnforcePermission to allowlist"
*frameworks/libs/modules-utils updated to 5a368d2c00f0e52257f977096cab69e0aa4b1580
Add @EnforcePermission annotation

Introduce a new annotation targeting system services. Similarly to
@RequiresPermission, it indicates the permission(s) required by a method
when called. As opposed to @RequiresPermission, the verification is
strictly enforced on the server side, based on the exact expression of
the annotation.

This Java annotation is automatically generated for AIDL interfaces that
use the corresponding AIDL @Enforce annotation. The intent is to declare
the required permission(s) when defining an interface and consistently
implement the associated verification.

Bug: 197828948
Test: atest --host aidl_unittests
Change-Id: I7097322fcd81b1e38e8931b757030e174b749d00

*system/tools/aidl updated to c6770e9708cac9a42938f14c002ba7de31839b1a
Translate AIDL permission annotations to Java

The @EnforcePermission and @RequiresNoPermission AIDL annotations will
generate an identical annotation on the Java binder methods. This
annotation will help surfacing to the service implementation that
authorization has already been verified.

Test: atest --host aidl_unittests
Bug: 197828948
Change-Id: I5bc2456d56c346c256891b2c96e88efeae129013