odrefresh: add flag of system server compiler filter
Similar to --zygote-kind, the flag allows the caller to specify system
server compiler filter, with higher precedence than the corresponding
system property (dalvik.vm.systemservercompilerfilter).

Since the property is used earlier (where the default value is used),
the test is updated to test explicitly with "speed" filter, instead of
checking the default behavior implicitly. Also, since we're not
expecting unknown filter, when an unknown value is given, the actual
code now crashes with UNREACHABLE().

Bug: 208269838
Test: atest art_standalone_odrefresh_tests
Test: atest ComposHostTestCases
Change-Id: I4f5335e872b245dc6156cddd944800120adc60da