liblog: Refactor log_wrap_test
The prior unit test reads MAIN buffer with ANDRID_LOG_WRAP flag.
But the reading thread wakes up when the MAIN buffer is full.
The MAIN buffer is usually filled less than 5 seconds.
For example, the cf_x86_64_phone-userdebug target.
So we switch to read SYSTEM buffer, which is stuffed much slower.

Bug: 197901557
Test: CtsLiblogTestCases
Change-Id: Ib53cc25469d2049f362a019c36bdd30f9e634b15