Autoupdate of 4 submodules
*art updated to 3ba3a77e4971b1ae411a18debc5e3507885c3b45
Update for MapInfo objects in frame data.

The frame data no longer contains map_XXX fields which represent
the map data. Now there is only a shared pointer to the MapInfo
object with which this frame is associated.

Bug: 120606663

Test: Compiles.
Change-Id: I1faf83a1023fe8344c16807599397ecceba2ccd2

*system/core updated to c54ddc2b72b1af1540cc2a8d58fa664183b6c163
Merge "Update for MapInfo objects in frame data."
*system/extras updated to 7c1ac352574621daf232d01d6f5960585f2829f9
Merge "Update for MapInfo objects in frame data."
*system/unwinding updated to b61aebb7f7afaad9b775a59dcd0ea6289fc4325a
Merge "Move MapInfo objects in to frame data."