base: normalize uid_t and pid_t across the codebase
We ended up in a state where different parts of the codebase
defined uid_t and pid_t (on Windows, where it's missing).
That was confusing. Instead this CL introduce a sys_types.h
header that takes care of that, without leaking further headers.
Also fixing the pid_t definition on Windows, which is supposed to
be a signed type.

Bug: 175152326
Change-Id: I7b95d7965548702185bdc49ec2b9ac2ad4abd5cd