Adds a way to invoke configs not local to Tradefed
This is a feature to enable a more streamlined configuration pipeline
for Tradefed that enables us to have more reasonable config parsing in the
future and to execute configs not necessarily local to the Tradefed
instance (ie downloaded from GCS).

Now includes a simpler file resolver and resolver loader since the
existing stuff is quite tied to resolving options in a configuration.

Bug: b/209688105
Test: WIP
Change-Id: I2def03f6656d7895b704e4a7a95fb9e8af4191b2