ANDROID: pvmfw: Receive image size through x2
Update the ABI with the VMM to expect the size of the image in register
x2 as it is otherwise impossible for pvmfw to deduce the size of the
image pointed to by x1 (as the ABI doesn't require the image itself to
contain its size e.g. through a standard header format).

Bug: 203505371
Test: -
Change-Id: Icd49aa2767f5ce71e0bcb5c627d0ef731056d89d