Roll-forward fix for bazel_apexer_wrapper_test.
An upstream change to apexer added a proto field to the default
apex_manifest.pb output, which fails in one of our relatively fragile
proto binary diff tests.

Since the test does not actually need to test for the contents of the
proto (but adjacent symlinked files), relax the test to an existence
test instead.

Fixes: 215129834
Test: b test //build/bazel/rules/apex:bazel_apexer_wrapper_test
Change-Id: Iaeceee53984a4ac9f6080e35d0680faa06517ffe