Support ase attribute handle assignment as dump ISO data
Some hci log file didn't contain the pairing process, the tool could not
get the correct codec paramter by parsing the packet automatically due
to the ASE control point attribute handle missing. Add the feature to
make it eaiser. The user can assign the ASE attibute handle used for
parsing. The tool could based on it to get the correct codec parameter
and generate the header for the decoder

Bug: 200748923
Bug: 150670922
Test: dump ISO data with ase handle assign on the log without pairing process
Change-Id: I67024709c37314a0f16e471368cfd9872c35f79b