Fixed the bug of missing attribute "sdm_log_path"
The bug of missing attribute "sdm_log_path" in stop_sdm_logger is
fixed.

Since stop_sdm_logger will be always run in the teardown procedure
this bug can be seen when SDM logger is never started because the
initialization of "sdm_log_path" is done only in start_sdm_logger.

Bug: None
Test: Yes, locally
Change-Id: Idc3d92475dee33fdd6411742e796ace58604a391