build libelf-free minimal libbpf for device
Add a device build of libbpf that includes only code with no libelf
dependency, and which can therefore be used from framework code.

bpf.c depends on a few symbols implemented in libbpf.c, so add a new
file with stub implementations for those symbols.

Also, the local linux/compiler.h and linux/types.h clobber the bionic
versions, with the result that __user, __force and __poll_t are
undefined. Since other bionic headers expect these to be defined, this
causes most of the files under src/ to fail to compile. Add a header
under android/ to ensure these are defined, and include it via
Android.bp to minimize Android-specific changes to upstream libbpf
code.

Test: m libbpf_minimal
Bug: 203823368
Signed-off-by: Connor O'Brien <connoro@google.com>
Change-Id: I326b31cac62bbd58905d25d45483cf968c00b971