Autoupdate of 2 submodules
*frameworks/base updated to 55d337b11b474af4a6d198e959cb38238bdde0ef
Merge "Add carrier config for EAP-AKA fast reauth feature"
*packages/services/Iwlan updated to 89789cdf33f29103316a806f9666a4d84d334415
[Iwlan] Support EAP-AKA fast reauth

Add support for the EAP fast reauth feature.
EpdgTunnelManager will cache the reauth ID in the Eapinfo if it's
available. And uses it to compose the Nai info and enable EapAkaOption
with next reauth ID in the next IKE tunnel setup.

The cached reauth ID will be cleared when the tunnel close
exceptionally.

Bug: 195961691
Test: Fast reauth function was verified with lab testing
Change-Id: I09ebe5b3cea6a20207ece1102d5a704f036bca27