Remove user-select defaulted to none for the whole page
Intent of "user-select: none" is to disallow the text selection on the
timeline. This CL moves the rule to ".panels" selector, achieving the
same effect but without needing to sprinkle "user-select: text" all
over the other CSS rules. Additional class "track-title" was added to
allow selecting name of the track group; this hasn't been done for
tracks because they implement drag'n'drop behaviour.

Change-Id: Ida0ab5a92b6a0d443474b60787ded0536b2d248d