Add toggleable attribute to manifest for run trace quick settings tile.
Added meta-data element to QsService element in the AndroidManifest
enabling the toggleable option for the run trace quick settings tile.

Bug: 215250733

Test: Manual. Prior to change the run trace quick settings tile
contained a > after the change the > is no longer there.  Tapping the run
trace button still starts and stops a trace with share trace popup still
showing when a trace is complete. Long pressing the run trace button
still opens the Traceur app.

Change-Id: Ia0f95535761d9a0453204a71ea2943eae2e011ab