Tag AWT windows as sRGB surfaces
OSX assumes windows are in the display's color space by default.
Java2D assumes its output in sRGB, which means the sRGB colors
are interpreted by the OS as being in the display's color space.
This means that unless the display is set to use the sRGB color
space, the final result on screen will look incorrect.

Issue: 208508346
Change-Id: Ie4c8dd2e41ba4f92db0c59e94825512ea93c6553