Autoupdate of 2 submodules
*system/bpf updated to 0ea4c6b242168dc7798bc0019cfcb73e4655281d
bpfloader: don't depend on symtab ordering to find function name

readCodeSections implicitly assumes that a function's name will be the
first symbol in its section to appear in the symbol table, but this is
not guaranteed. In particular, when compiling with -g the section name
can appear before it, resulting in a failure to find the bpf_prog_def
information for that section's program.

Add a check for the STT_FUNC symbol type to narrow our search to the
function name symbol we actually want.

Also, add a test to check that a program with a minimum kernel version
of infinity will not be loaded, to verify that bpfloader successfully
finds bpf_prog_def information.

Test: boot & confirm progs all have correct owners & permissions
Test: build time_in_state.c with -g and confirm owner & permissions
set correctly
Test: libbpf_load_test now passes iff this fix is applied
Signed-off-by: Connor O'Brien <connoro@google.com>
Change-Id: I31db0c50d3adeed86ee5810b1a0fb972f56fdb47

*system/bpfprogs updated to 3e438a069d266272a2313328f247f41b30181ddb
bpf_load_tp_prog: add a second prog to test kernel version checking

Add a sched_wakeup program with a minimum kernel version of
infinity, to allow testing that bpf_prog_def information is being
found & enforced correctly.

Bug: 203823368
Test: libbpf_load_test
Signed-off-by: Connor O'Brien <connoro@google.com>
Change-Id: I9ce3da1e4b0d70e4e9aa8838b404f611788f9911