Set exclude_kotlinc_generated_files to false explicitly since kotlin reflection is used in NetworkStack test cases kotlin code.
Property "exclude_kotlinc_generated_files" is recently added to exclue some kotlinc generated files that are not needed if kotlin reflection is not used. Currently it defaults to false since there are test failures(b/213966211) which is in NetworkStackServiceTest.testDumpVersion that uses kotlin reflection and fails if exclude_kotlinc_generated_files defaults to true. By setting explicitly  exclude_kotlinc_generated_files to false on tests like NetworkStackServiceTest.testDumpVersion, we will be able to finally default this property to true so the kotlinc generated files can be removed for most modules that don't need them,

Bug: 204888276
Bug: 213966211
Test: atest ConnectivityCoverageTests:com.android.server.NetworkStackServiceTest

Change-Id: I370ce598195c0c1d3d0e996cc31368484ff15509