Switching the WorkTask to a unique registration
This change fixes a bug in which multiple instances of the
PeriodicProvisioner Task could be scheduled, due to the enqueue method
on boot not specifying a unique task. The task replacement policy is set
such that a new task will always run on boot and then replace the old
task in the scheduler.

Test: "adb shell dumpsys jobscheduler" and check for the job
Change-Id: I6e05d0057fa9be7cfea25a4cc3424ce90d7ddbaf