Autoupdate of 2 submodules
*packages/modules/Virtualization updated to 49e9fbd348b389ca6dd49b9c410838b3bf3dab27
Merge "Pass isFactory bit from host to Microdroid via payload metadata"
*system/apex updated to 7dd5b1e7000d1eee07ec702849d73a281da11f3b
isFactory in VM mirrors isFactory in host

Previously, all APEXes that are activated in the VM were considered as
pre-installed (i.e. isFactory=true), regardless of whether an APEX that
is imported from the host is actually a non-factory version there.

This has caused problem to the isolated compilation usecase where
isFactory bit of an APEX is encoded in the compiled artifact and matched
against the bit on the host side.

With this CL, isFactory is the same across VM and the host. This is done
by adding non-factory APEXes (which can be identified by the added field
in the payload metdata) to data_store, rather than pre_installed_store.

There however is a small caveat. Public key of an APEX is found not only
from pre-installed ones, but also from data ones, IF the APEX is backed
by a block device. This is needed because, unlike the host-side case, we
don't have both a pre-installed and a data version of the same APEX in
the VM. (We could do that, but that would be just waste of memory and
cpu cycle).

Bug: 210998761
Test: atest ApexTestCases

Change-Id: I8c82be4335632a62c20b892dbc13a3ec3036c937