Add annotations zip, and system and module-lib APIs
At the moment they are all empty but they are needed to ensure that
the prebuilts provide the same outputs as the source modules.

Updated the files by first running this command:
   TARGET_BUILD_APPS=com.android.conscrypt \
       packages/modules/common/build/mainline_modules_sdks.sh

Then unzip the generated snapshot into the current directory:
   cd prebuilts/module_sdk/conscrypt/current
   unzip $ANDROID_BUILD_TOP/out/dist/mainline-sdks/current/com.android.conscrypt/sdk/conscrypt-module-sdk-current.zip

Then keep only those files and Android.bp changes related to the
addition of annotations and APIs.

Bug: 216435117
Test: m nothing
      - TH will test the prebuilts in master-art
Change-Id: I8eaaa9a5f9644c8f32e39ea108dc2616ae345920