Autoupdate of 2 submodules
*art updated to 6740fccc2dcbd868d2144b206806a880ae486ca8
Don't do force-compile in CompOS mode

This was originally done to ensure that CompOS didn't reuse old
untrusted artifacts. But that won't happen - CompOS always runs with
the output directory initially empty; authfs doesn't provide a
mechanism to have a writable directory with untrusted files in it
(b/203630168).

And it is harmful - e.g. if ART has not been updated the force-compile
will generate a boot.art, which will then be deleted on reboot, and as
a consequence the other artifacts won't be loaded into system server
etc.

Bug: 211458160
Test: atest OdrefreshHostTest#verifyCompilationOsMode
Test: manual
Change-Id: I888e4b13dd29e298f6635569d5f06d36bc672fc8

*packages/modules/Virtualization updated to 2d2e4dba53eb425c7f4914234b19aced6f21e057
Use --force-compile for tests

The test uses --force-compile outside the VM, so we also need to
inside the VM. But we want --compile for the normal case of
compilation, so we only produce exactly the required artifacts.

That means we have to pass whether we are testing or not from the host
to guest, so lots of plumbing is needed.

(Also figured out how to solve a problem with type deduction that the
me of a few months ago was baffled by.)

Bug: 211458160
Test: atest ComposTestCase
Change-Id: Ie7e36bc3fab40744e9f9ea5e5928126899d9060d