Limit memory used by filtering into file
Before this commit, the code would:
* read all the data from the buffers
* filter all the data.
* write it to the file.

Reading the data would only allocate a small amount of memory (for the
Slice pointers and trusted data).

Filtering would allocate a lot of memory instead, which would exhaust
the watchdog.

This commits changes the code to read, filter and write smaller chunks of
data, to reduce memory allocation.

This commits also fixes a small problem: ReadBuffers() should not
generate a "read complete" lifetime event, unless it read all the data
from the buffers. This was not a big problem (until now) because:
* For the write_into_file cases, ReadBuffers() always read all available
  data. This assumption is changing with this commit, so this has to be
  fixed now.
* For the send to consumer case, we didn't have many consumers that
  relied on periodic flushes (perfetto_cmd first disables tracing and
  then reads all the data). In the future we might have a consumer that
  uses periodic flushes though, so it's nice to fix the bug now.

Tested: Recorded traces with filtering with and without write_into_file
and huge buffers

Bug: 195145848
Change-Id: I1ab412eb8e1bd3566f8679a318a9868aedea892b