Fix flush race triggered from TracingServiceImpl::StartTracing
This specific problem has been discovered when running relatively short
chromium startup tracing sessions. On some of test devices we often hit
the following DCHECK in chromium codebase.

This specific race manifests itself by triggering the following DCHECK
  [3882:3891:0126/115600.790166:3682185893:FATAL:trace_event_data_source.cc(1200)]
    Check failed: TraceLog::GetInstance()->IsEnabled().
  #0 0x0000036f50d8 base::debug::CollectStackTrace()
  #1 0x0000036703b4 base::debug::StackTrace::StackTrace()
  #2 0x00000367faa4 logging::LogMessage::~LogMessage()
  #3 0x0000036802b0 logging::LogMessage::~LogMessage()
  #4 0x00000366b66a logging::CheckError::~CheckError()
  #5 0x0000039d6ec2 tracing::TraceEventDataSource::Flush()
  #6 0x0000039d3b6c tracing::ProducerClient::Flush()
  #7 0x0000039e2d28 tracing::mojom::ProducerClientStubDispatch::Accept()

When it happens not only TraceLog instance has been disabled. The
TraceEventDataSource::IsEnabled method also returns false. Basically
somebody tries to flush data source after it has been disabled.

The root cause for this can be traced back to perfetto's
TracingServiceImpl::StartTracing method. This method among other things
post a delayed task to call FlushAndDisableTracing if duration_ms config
key is non zero. Unfortunately this delayed callback does not take into
account that tracing could already be disabled when its invoked. This
can happen if somebody already called TracingServiceImpl::DisableTracing,
as is often the case in application shutdown process. If the delayed
callback is invoked in the middle of the shutdown procedure the extra
flush can potentially prompt the code to reach a state its not expected
to reach. In case of chromium this can trigger TraceEventDataSource
illustrated on the provided stack trace. This DCHECK is not simple to
reproduce since the TraceEventDataSource code runs in a different
process which often is also shutting down.

Since the code generally does not expect to call flush when tracing is
disabled, lets just make sure TracingServiceImpl::FLush does nothing if
tracing has has not been started.

Change-Id: Ie7944b004cfdff216199217da918ba614f3ff552