Add handling for toolchain/rustc/Android.mk
Some crates in the Rust source contain Android.mk files.  While these
files are stripped from the generated prebuilts we leave them in the
source repository to keep the files as close to upstream as possible.

This can cause a problem in the aosp/master-plus-rust repository when
Soong goes to configure itself as it can/will find incorrect definitions
for various modules.  To prevent Soong from scanning the toolchain/rustc
directory a blank Android.mk file has been added to toolchain/rustc.

This CL updates the build system to correctly handle the guard
Android.mk file when updating the source files in toochain/rustc as well
as when generating the prebuilts.

Bug: 166837771
Change-Id: Ic6a648bd26b75288337ec7ad3dc6e59d393f04c7