ANDROID: Add env verified_import command and use it
Implement "env verified_import" to enable loading U-Boot environment
that is supplementary to the default environment (use '-d' to replace
the default environment).

The environment keys are only loaded if the partition is in the vbmeta
chain and verification is successful. If either the partition is not in
the chain, or the partition data has been tampered with, the keys will
not be imported. This implements a secure way to augment the default
U-Boot environment.

This code is almost entirely the same as aosp/1964846, but it is not
limited to running at env initialization time and can work with any
interface, device or named partition, in any format supported by the
U-Boot avb code.

Bug: 216304872
Change-Id: I1e1728c348837f1d3db87b0f35c8bb6244386ca7