Adding a builder pattern for starting Microdroid
Instead of using fixed parameters for starting a Microdroid in Android
device, adding a builder pattern so that if in future we need to support
more flags, we can do it without breaking exisiting code. Also modifyig
the design so that it is easy to support multiple microdroid devices in
future if we need to.

Bug: 182478337
Bug: 216664481
Test: abtd, "tradefed.sh run host --class com.android.tradefed.device.MicrodroidFuncTest"
Change-Id: Ifac2d5559d6468393ddf44ffc7aebf06ee830fd2