Default to signature format v4 in tests
This sets the default for DriverTest tests to use signature format v4.
To ease bumping the version in the future, DriverTest now prepends a
signature format header to expected signature files based on the
supplied format if one wasn't supplied.

Existing tests have been fixed by specifying an explicit format version
in the test. For most tests, this was V2, except for tests which had
specified V3 in the format version header in the test. These tests will
be modernized if appropriate in a subsequent CL.

Test: Updated existing tests
Bug: 190515646
Change-Id: I17bbe4bb2d56a9a764309143d87e91168caead30