perf events: expand set of builtin counter enums for sw and hw events
The proto now covers the software and hardware enums present in the 5.13
linux user api header. This is still a fairly narrow subset of what's
usually available on a system, since we're not parsing dynamic PMUs, nor
device-specific counters.

There's an argument for having a single header, or a script, to glue
together all of the proto<->name<->structs conversions, but for now it
still seems reasonable to just have a couple of big switch cases in
their relevant files.

I started with the proto enum, and generated the code with find-replace.
So if there are any leftover copypaste errors, they'd be present
everywhere. The fact that it builds at least means that the enum names
and the uapi names are consistent.

Change-Id: I148b7c51e6cca775568762324201a96f1afa9afa