Fix lint errors in DhcpResults.
Most of this CL is just cosmetic: add docstrings, fix style
issues, etc.

There is one substantive change: the class now has a non-default
hashCode method. This change is correct because the class
overrides equals, and hashCode should always be overridden if
equals() is overridden. This should not change any behaviour
because DhcpResults is never put into any containers:

https://cs.android.com/search?q=usage:%5CbDhcpResults%5Cb%20-file:aidl_api%20case:yes%20file:packages%2Fmodules%2FNetworkStack%20-file:Test%5B.%5Djava$

Test: atest NetworkStackNextTests NetworkStackNextIntegrationTests
Test: device connects to wifi, logs contain readable DhcpResults
Change-Id: I36ab7e30f562dbb9703665fb9e51bdb94fc5d109