Fixup Fuchsia's waiting for no wifi connections
If there are already no client connections, no update will be generated
when the device is requested to disconnect (e.g. by unsaving a network).

In a test like HiddenNetworksTest:test_auto_connect_hidden_on_save,
this causes a timeout when we await this update.

Fixup by forcing the reading of a new update each time.

Bug: None
Test: act.py -c ... -tc HiddenNetworksTest
Change-Id: I24cdf1f3b0d1f4017f5310767b8cf7f146fc9c39