Use a handler to process screen record file
Previously screen record file is returned as a file object. Later after the tests are migrated to junit, the run job started to throw assertion exception when test failed. This causes the method to exit without returning the recorded video file. This change make the method use a handler to process the video file so that the caller can still access the video file when an exception is thrown.

Test: atest csuite-harness-tests
Fix: 217090486
Change-Id: Ic490dfc12009b40952a251f1cac8e6a42a998949