Autoupdate of 4 submodules
*external/error_prone updated to c2676a8fd19b6e7bbb4d7ec56dfdd1e187e520d5
Update errorprone to 2.11.0

- ./update.sh 2.11.0 9+181-r4173-1 3.15.0
- Changed 2.10.0 -> 2.11.0 in Android.bp
- Removed -Xep:RemovedInJDK11:WARN as it no longer exists
- Add -Xep:EqualNull:WARN because there are a lot of instances of it
- Removed commented out flags

Bug: 190944875
Test: Presubmits
Change-Id: Id3918e24aa45e2b986cbe6f1e866fdd5d3154331

*external/javaparser updated to 2686dc2a30241691758beb45ae077e646dcbb050
Demote UnicodeInCode error to warning

Errorprone is being updated, and adds a check
that errors out when there's unicode in java code.
Javaparser has a method called
thereIsSomeJSuchThatβequalAlphaJ, so demote the
error to a warning.

Also remove the OWNERS file because the only
two people in it left Google.

Bug: 190944875
Test: Presubmits
Change-Id: I2ceb5b4c18c16354959b23f63ed6dda7ea85a8c0

*external/tagsoup updated to 04aba7cd551194dbdb388a0127d81aa13a859ec7
Merge "Demote UnicodeInCode error to warning"
*libcore updated to 9d0bea3006f1a5844368491de5fb12e14d44ed87
Merge "Demote errorprone errors to warnings"