Quadruple monitor table size in debug build
The monitor table in Android debug builds was intentionally very small,
so that we could test expansion to more than a single level. This has
become problematic since debug builds are increasingly used to run
frameworks tests.

As a compromise, increase the size by a factor of 4. This is still a
factor of 64 smaller than production builds.

Bug: 217607468
Test: Build boot AOSP, Treehugger
Change-Id: I54116bac2b9e6743bac0feccdde19648351ae04d