Silent failures of MMTEL callbacks causing SMS to fail
We are silently failing if an app tries to register a callback
before telephony has processed the indication to bring up IMS on
that subscription. This is causing apps to think that registration
of the callback completed successfully when it actually didnt.

For now, revert the subId check and make it a warning as well as
beef up logging around this so we can see this condition happening
better in logs.

The next step is to go through all of the register* methods for
IMS and ensure they send back proper Exceptions when they fail.

Bug: 216845706
Test: Send SMS on P6 before and after fix
Change-Id: I1da1ade025ca6a1a897e3a59ec9585a4726fc9b1
Merged-In: I1da1ade025ca6a1a897e3a59ec9585a4726fc9b1