odsign_e2e_tests: Use the current ART module.
Before this change, the test uses a test apex. However, it doesn't work
in CTS because the key used to sign the test apex in CTS doesn't match
the key that the device expects. After this change, the test uses the
current active ART module on device, which is always valid.

Bug: 215552478
Bug: 184143265
Test: Flash a prod system image and run odsign_e2e_tests build from
  AOSP.
Change-Id: Ibbc9614cfab12feca223fa44ce7550690c51a79d