Add linker scripts for embedded linker to sysroot
Add linker scripts that trick clang into produce a binary with the
embedded dynamic linker.  Scrt1.o is now a linker script that includes
the original Scrt1.o and libc_musl_linker_object.o that contains the
linker and the trampoline, as well as the linker.script that puts
the linker in the right place.  libc.so is a linker script that forwards
to libc_musl.so.

Bug: 216192129
Test: manual
Change-Id: Ibbc749460fb856c57deda87b9ed3c252b67a708f